GROW-77 refactor: Summary, Summary Card 서버 컴포넌트로 수정 및 오늘의 review 디자인 추가
